Migrated project from React.js to Next.js
