--- index.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <link rel="stylesheet" href="style.css" />
-    <meta charset="UTF-8" />
-    <title>CA-PHE</title>
-  </head>
-
-  <body>
-    <header>
-      <div>
-        <img
-          class="logo"
-          src="images/ca phe mock client profile transparent.png"
-        />
-      </div>
-      <div>
-        <nav id="navbar">
-          <ul>
-            <li><a href=""> Menu </a></li>
-            <li><a href=""> Seasonal Dishes </a></li>
-            <li><a href=""> Inside the Cafe </a></li>
-            <li><a href=""> Delivery FAQ </a></li>
-            <li><a href=""> About Us </a></li>
-            <li><a href=""> Our Advocacy </a></li>
-          </ul>
-        </nav>
-      </div>
-    </header>
-  </body>
-</html>
