+++ pages/FAQ.js
