+++ pages/_app.js
@@ -1,0 +1,6 @@
+import Head from "next/head";
+import "../styles/styles.css";
+
+export default function MyApp({ Component, pageProps }) {
+    return <Component {...pageProps} />
+}
