--- src/App.css
@@ -1,78 +1,0 @@
-@charset "UTF-8";
-/* CSS Document */
-
-:root {
-    --logo-color1: rgb(167, 149, 131);
-    --background-color1: rgb(250, 249, 247);
-    --navbar-bg: rgb(239, 234, 228);
-    --text-color1: rgb(51, 50, 39);
-}
-
-html {
-    box-sizing: border-box;
-}
-
-body {
-    margin: 0;
-    padding: 0;
-    overflow: hidden;
-    width: 100%;
-    height: 100vh;
-    background-color: var(--background-color1);
-}
-
-header {
-    display: flex;
-    flex-direction: row;
-    margin: auto;
-    position: absolute;
-    width: 100%;
-}
-
-    header div {
-        width: 100%;
-    }
-
-#navbar {
-    background-color: var(--navbar-bg);
-    color: var(--text-color1);
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-    #navbar ul {
-        list-style: none;
-        margin: 0;
-        text-align: center;
-        display: flex;
-        flex-direction: column;
-        flex-wrap: nowrap;
-        font-size: 20px;
-        padding: 0;
-        position: absolute;
-        right: 50px;
-    }
-
-li a {
-    text-decoration: none;
-    color: var(--text-color1);
-}
-
-li {
-    border-color: black;
-    border-style: solid;
-    border-width: 1.5px;
-    border-radius: 5px;
-    margin: 10px;
-    padding: 1rem;
-    padding-inline-start: 4em;
-    padding-inline-end: 4rem;
-    width: 15rem;
-    display: block;
-}
-
-img {
-    max-width: 100%;
-    flex-direction: column;
-}
