--- src/pages/LandingPage.css
@@ -1,78 +1,0 @@
-:root {
-    --logo-color1: rgb(167, 149, 131);
-    --background-color1: rgb(250, 249, 247);
-    --navbar-bg: rgb(239, 234, 228);
-    --text-color1: rgb(51, 50, 39);
-}
-
-body {
-    margin: 0;
-    padding: 0;
-    overflow: hidden;
-    width: 100%;
-    height: 100%;
-    background-color: var(--background-color1);
-}
-
-.LandingPage {
-    
-}
-
-.LandingPage__Logo {
-    max-width: 50%;
-    flex-direction: column;
-}
-
-.LandingPage__nav-links {
-    background-color: var(--navbar-bg);
-    background-size: contain;
-    color: var(--text-color1);
-    list-style: none;
-    margin: 0;
-    align-items: center;
-    justify-items: center;
-    display: flex;
-    flex-direction: column;
-    flex-wrap: nowrap;
-    font-size: 20px;
-    padding: 0;
-    position: absolute;
-    right: 50px;
-    top: 20px;
-    height: 57%;
-}
-
-    
-    .LandingPage__nav-links a {
-        text-decoration: none;
-    }
-
-    .LandingPage__nav-links button {
-        background-color: var(--navbar-bg);
-        border-color: black;
-        border-style: solid;
-        border-width: 1.5px;
-        border-radius: 5px;
-        font-size: 20px;
-        margin: 10px;
-        padding: 1rem;
-        padding-inline-start: 4em;
-        padding-inline-end: 4rem;
-        width: 25rem;
-        display: block;
-    }
-
-    .LandingPage__nav-links button:hover,
-    .LandingPage__nav-links button:active {
-        background: #808080;
-        color: white;
-    }
-
-.LandingPage__nav-links-socials img {
-    width: 2.5rem;
-    height: 2.5rem;
-    padding: 1rem;
-    padding-inline-start: 1.5rem;
-    padding-inline-end: 1.5rem;
-    display: inline-flex;
-}
