+++ styles/styles.css
@@ -15,9 +15,9 @@ html {
 body {
   margin: 0;
   padding: 0;
-  overflow: hidden;
-  width: 100%;
-  height: 100vh;
+  overflow: auto;
+  width: auto;
+  height: auto;
   background-color: var(--background-color1);
 }
